+++ src/templates/artifacts/StartupView.groovy
@@ -8,13 +8,12 @@ import java.awt.event.*
 
 actions {
 <%
-   g.domainClasses.each { String name, def domainClass ->
+   g.domainClasses.findAll{ k,v -> v.entity }.each { String name, def domainClass ->
 	   def prop = griffon.util.GriffonNameUtils.getPropertyName(name)
 	   def natural = griffon.util.GriffonNameUtils.getNaturalName(name)
 	   out << "\taction(id: '$prop', name: '$natural', actionCommandKey: '$prop', closure: controller.switchPage)\n"
    }
-%>
-}
+%>}
 
 application(id: 'mainFrame',
   title: app.config.application.title,
@@ -27,12 +26,11 @@ application(id: 'mainFrame',
 	toolBar(constraints: BorderLayout.PAGE_START, floatable: false) {
 		buttonGroup(id: 'buttons')
 <%
-	g.domainClasses.each { String name, def domainClass ->
+	g.domainClasses.findAll{ k,v -> v.entity }.each { String name, def domainClass ->
 		def prop = griffon.util.GriffonNameUtils.getPropertyName(name)
 		out << "\t\ttoggleButton(buttonGroup: buttons, action: $prop, verticalTextPosition: SwingConstants.BOTTOM, horizontalTextPosition: SwingConstants.CENTER)\n"
 	}
-%>
-	}
+%>    }
 
 	panel(id: "mainPanel", constraints: BorderLayout.CENTER) {
 		borderLayout()